Update mullvad_speed_test.py
A number of changes have been made: rapid testing of the Mullvad global network of servers to determine average response time, localisation of the test, procedure for cancelling the current test and moving on to another if the server is unreachable, improved graphical interaction, structuring of results tables, etc. Have fun!